@@ -7,7 +7,7 @@ name = "ai-human-inventory-game"
 version = "1.0.0"
 description = "Fullstack web application for exploring human-AI collaboration in inventory management"
 readme = "README.md"
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 license = {text = "MIT"}
 authors = [
     {name="Tianyi Peng", email="tianyipeng@gmail.com"},
